solution

+++ prueba_tecnica_linktic/src/main/java/com/example/prueba_tecnica_linktic/DTOS/response/CalculePrimaTotalDTO.java
@@ -33,7 +33,7 @@ public class CalculePrimaTotalDTO {
     @JsonProperty("nro_identificacion")
      String nroIdentificacion;
     @JsonProperty("valor_asegurado")
-     Float valorAsegurado;
+     Integer valorAsegurado;
     @JsonProperty("liquidaci\u00f3n")
      List<LiquidaciN> liquidaciN;
     @JsonProperty("Valor_total")
